+++ src/oci/resource_manager/models/create_stack_details.py
@@ -37,6 +37,10 @@ class CreateStackDetails(object):
             The value to assign to the variables property of this CreateStackDetails.
         :type variables: dict(str, str)
 
+        :param terraform_version:
+            The value to assign to the terraform_version property of this CreateStackDetails.
+        :type terraform_version: str
+
         :param freeform_tags:
             The value to assign to the freeform_tags property of this CreateStackDetails.
         :type freeform_tags: dict(str, str)
@@ -52,6 +56,7 @@ class CreateStackDetails(object):
             'description': 'str',
             'config_source': 'CreateConfigSourceDetails',
             'variables': 'dict(str, str)',
+            'terraform_version': 'str',
             'freeform_tags': 'dict(str, str)',
             'defined_tags': 'dict(str, dict(str, object))'
         }
@@ -62,6 +67,7 @@ class CreateStackDetails(object):
             'description': 'description',
             'config_source': 'configSource',
             'variables': 'variables',
+            'terraform_version': 'terraformVersion',
             'freeform_tags': 'freeformTags',
             'defined_tags': 'definedTags'
         }
@@ -71,6 +77,7 @@ class CreateStackDetails(object):
         self._description = None
         self._config_source = None
         self._variables = None
+        self._terraform_version = None
         self._freeform_tags = None
         self._defined_tags = None
 
@@ -195,6 +202,30 @@ class CreateStackDetails(object):
         :type: dict(str, str)
         """
         self._variables = variables
+
+    @property
+    def terraform_version(self):
+        """
+        Gets the terraform_version of this CreateStackDetails.
+        The stack's Terraform version
+
+
+        :return: The terraform_version of this CreateStackDetails.
+        :rtype: str
+        """
+        return self._terraform_version
+
+    @terraform_version.setter
+    def terraform_version(self, terraform_version):
+        """
+        Sets the terraform_version of this CreateStackDetails.
+        The stack's Terraform version
+
+
+        :param terraform_version: The terraform_version of this CreateStackDetails.
+        :type: str
+        """
+        self._terraform_version = terraform_version
 
     @property
     def freeform_tags(self):